backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.8.0",
    "djangorestframework>=3.16.1",
]
[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]
